refactor: Move tests to the main test directory
#4710: Improve project hierarchies
