+++ akvo/rsr/tests/views/test_indicator_period_label.py
